backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.6
python-socketio==5.10.0
google-generativeai==0.5.0
python-dotenv==1.0.0
PyPDF2==3.0.1
gevent==23.9.1
langgraph==0.0.21
langchain==0.1.0
langchain-google-genai==0.0.5
opencv-python==4.8.0.76
deepface==0.0.75

Pillow==10.0.0
werkzeug==3.0.0
email-validator==2.1.0
flask-limiter==3.5.0
flask-swagger-ui==5.0.0
lembic==1.13.0a
pytest==7.4.0
pytest-flask==1.3.0
pytest-cov==4.1.0
openai>=1.35.0
gunicorn==21.2.0
gevent-websocket==0.10.1
